refactor: move all service related errors to service file
